CircleCI config, deploy to dev.jonogmarteinn.is when on dev branch
Updated the CircleCI config.

* Updated comments regarding the homebrew solution to install aws-cli. CircleCI now offers Orbs and there is an official aws-cli orb so this step is no longer needed. However Orbs require to enable pipeliens. There were some warnings about enabling that option and the warning that if you want to go back you have to get in touch with CircleCI customer support. I will have to look into and try Orbs at a later time. For now I will just stick with this installation script because I know it works reliably.

Added a new step that creates a robots.txt file. If we are on the master branch (about to deploy to production) we set the file to indicate to search engines that they are free to index all of the site (jonogmarteinn.is). If we are on the dev branch (dev.jonogmarteinn.is) the content signals to search engines that we do not wish for it to be listed in search results. It's a development environment to test things out.

Added a conditional check to the deploy step. I added the AWS resources for dev.jonogmarteinn.is. Now when we push on the git branch "dev" we automatically deploy to dev.jonogmarteinn.is.