Use red as a theme colour instead of the default blue
The theme colour of the Jón og Marteinn logo is red with hex value #cf1717. Added a new CSS file which overwrites the theme colours from Bootstrap and the Landing Page theme. It's not really an ideal solution. The proper way to do it would be to use the Landing Page SASS files and regenerate the vendor assets. I tried it but couldn't get it to work. The docs were very limited for the theme. I'm pretty sure it would also involve quite a bit of setup. At this stage I'm more interested in prototyping. Beside, this change may be all that is needed.