Made the email white and made the email wrap on smaller devices
Made the email white for now instead of the default red. On mobile devices the contrast with the current water background was poor. I would ideally like the email to be red to indicate it's clickable. However as a quick fix measure I just make it white.

Added the &#8203; right before the @ symbol in the email. This makes it such that the email can wrap like this on smaller phones:

jonogmarteinn
@jonogmarteinn.is

Before doing this the site looked all banged up on iPhone because how long the email is and how big I want it to be displayed.