Added portrait images of Jón Hafdal and Marteinn
Got some images from the photographer. The image of Jón is from the image IMG_0013.jpg and the portrait of Marteinn is from the image IMG_0040.jpg. I cropped to fit the head area then gradually faded the shadow out. Saved the downsized the images by about 33% and saved them as 99% JPG quality.

To note is that the images are not the exact same size. I'm not sure if that's going to be an issue or not. The reason why I didn't want to enforce the same size was to avoid downsampling twice. My thinking is that the size won't really matter anyway because the browser resizes them anyway to respond to browser size.

Each image is roughly 100KB. The image