Make images resize on the center instead from the top by default
The theme we use use background images for the showcase element. When we would resize the browser window the background position was by default top (in conjunction with "background-size: cover;"). That meant that as the image repositioned it prioritised the content on the top. Explicitly specifying center makes the center of the image more prominent and things scale better.