Added an image of Jón and Marteinn standing next to each other
Added a PNG image with a transparent background of Jón and Marteinn standing next to each other (full-body). Adjusted paddings to make things look a bit nicer.

Added background-position: center for the splash image. Similar change like in the previous commit. I think it looks nicer when the browser is scaled. I tried first with the value of "bottom" but didn't like it as much. The detail on the bottom was a bit more distracting. With center we display more of the blur distortion of the image.