Commented out all placeholder lorem ipsum text & social icons
I showed the site to Jón Hafdal last night and he was pleased with the work so far. I am hiding all placeholder text and non-functioning elements on the site so I can prepare to start using it instead on production. That is, the new site is now way better than what's on production (logo, subtitle, email, names and phone numbers).

There is more work to be done before the site can go public but this is a step in that direction.