Made text in footer larger and with better contrast,
Google Chrome Lighthouse audit gave the following penalty warning with -8 points for accessibility:

```
Background and foreground colors do not have a sufficient contrast ratio. 

Low-contrast text is difficult or impossible for many users to read. Learn more.
Failing Elements
p.text-muted.small.mb-4.mb-lg-0
```

The change makes the text both bigger and in a higher contrast. I think the text is now black (or dark grey) on a white background instead of the muted light grey on a white background.