Fixed a bug where profile pictures would overlap
I'm not sure why a max-width was being specified on profile images. By removing it the image simply fills out all it's available space which is just perfect. Then I can more easily adjust width with Bootstrap columns. I may be misunderstanding something but this fixed the bug and seems to work in all resolutions.