Made h2 subtitle on the splash screen a bit more responsive
Before this change the subtitle "alhliða málningarþjónusta" was pretty huge on small mobile devices. This change makes it so that the text appears more like a subtitle from the Jón og Marteinn lógo. Taking up aproximetly 80-90% of the width of the logo. It's not exact and there might be a better way to implement this, but this seems to have got the job done for now.