Added a new .heading and .overlay CSS classes
These classes are basically a copy-paste of the call-to-action CSS classes from the Landing Page theme. The only modification is a smaller top and bottom padding.

Essentially this is an experiment to have style <h2> headings in a way that creates a separation in the page. The background image is that of the closeup coarse wall.