Added new section to the page: Verkefni. Added project photos.
Added a new project section to the page below the types of services provided.

Stylistically it is the same as the types of service section above, that is an alternating 2x4 grid with photo on one side then text on the other, then switched for the row below (on desktop). I initially tried a different UI layout to break up the flow but I found this continuation to be pretty simple and elegant.

The drawback to this approach is that it becomes a bit harder to see become aware when scrolling through the page that there is new or different type of content. To address that I added "Þjónustur" and "Verkefni" <h2> headings and styled them with a dark grey background to create some separation. I'm not totally happy with how it looks. For example I think those headings take up too much space on mobile. But my plan is to refine that later on if I feel strongly about it.

I had a call with Jón to get some details about the projects that they have done. Apparently they have more projects about to be completed and I noted those down in a comment below the projects.

Added some photos that represent each project. 

There were more photos taken for each project. For now in this initial implementation we will just have one photo, heading and small description text of each project. But later I am deliberating how it would be best to achieve showing multiple photos for each project. A image slider comes to mind. Maybe a modal. Even further down the line I might get into a specific jonogmarteinn.is/verkefni, jonogmargeinn.is/verkefni/lindex route system. But to do that I will most likely have to expand into a static site generator. I would like to avoid that for as long as I can. The simiplicity of just having one HTML file and some CSS is really nice. No JavaScript either (that I write).