Bugfix: added no-gutters to section headings to remove horizontal scroll
After I added the "Þjónustur" and "Verkefni" headings the site could be horizontally scrolled. It was quite annoying when scrolling down on a mobile since you could scroll slightly to the right. Now when the site is scrolled down on a mobile the site goes just straight down and it doesn't have any horizontal scrollbar.