Added sliding images for the project Sólvangur
Added a Bootstrap carousel that slides through more photos from the Sólvangur project.

The carousel markup is mostly stock Bootstrap but with a bit of a twist from Start Bootstrap. I used the snipped from here:

https://startbootstrap.com/snippets/half-slider/

The images are a JPG export in 16.67% of the original size and saved with 81% compression.

See:
https://getbootstrap.com/docs/4.1/components/carousel/