Various CircleCI updates: updated config version, next-gen images
Updated to the latest version of the CircleCI config, 2.1. With it we can run pipelines that we don't actually need 😂

CircleCI is moving towards new Docker images that run builds and deployments. CircleCI claims faster spin-up times and improved reliability and stability. Neither has really been an issue for this site but I'm updating to it regardless because I suspect that the old circleci/node (and other) images will be deprecated in the future. It should be noted that according to Docker Hub (at the time of writing) the cimg/node image is still in beta.

https://hub.docker.com/r/cimg/node  

https://circleci.com/docs/2.0/circleci-images/#next-generation-convenience-images