Added Gatsby v2 static site generator as a dependency
Gatsby is a popular React static site generator. It's pretty powerful and configurable but I expect only to use a subset of the features.

I considered using Charge or Parcel instead. But since I have more experience with Gatsby from my personal blog I am going to stick with it. Charge looks really interesting but is probably more suitable to a different project. For now.

https://www.gatsbyjs.org/
https://charge.js.org/
https://parceljs.org/