Relocated all images and CSS files to the src/ folder
As far I can tell Gatsby expects all source files to be in the src/ folder.

All images were relocated from the public/img/ folder to src/images/ folder.

The two CSS files were relocated from public/css/ to src/ directly because there are only two of them. I may choose to relocated them to a styles or a CSS subfolder in the future but for now this will do.