Refactored public/index.html to become src/pages/index.js
Relocated the main index.html file to become index.js in the src/pages/ folder. I do this to retain the Git history that file has accumulated because it has been the main file for the site.

The page will be adapted to React and JSX in the next commit so as of this commit the file index.js does not work since it contains raw HTML contents.