Initial conversion of HTML to JSX for front page
A rough conversion of the HTML file to a React component. The main changes are that class attributes have been changed from "class" to "className". Images are imported through the packager for now. At a later stage I may choose to avail of the image compression infrastructure that Gatsby provides but for now I am mostly focused on bringing feature parity to the new Gatsby site to the old static Index.html file.

Things that have not been implemented yet:

* Proper <head> content, like <title>, description, viewport, etc.

* Font Awesome and Simple Line icons. Google font.

* Favicon. I removed the emoji favicon code by Kristján because I plan on using a favicon from the Jón & Marteinn logo.

* The image carousels on projects.

* The Instagram widget on the bottom of the page.

* There may be other things that I am forgetting.