Added public/ to .gitignore and moved vendor assets to src/
Moved the vendor assets to the src/ folder as a backup. At this stage I am not sure what if any of it I will use directly. So the plan is to delete the files after I have made use of them properly. I am guessing that most if not all are available through NPM.

Added the public/ folder to .gitignore because Gatsby uses the public/ folder as an export from builds. This means the public/ folder is no longer tracked in the repository.