Enabled the LightWidget that displays the Instagram photos
LightWidget is used to display the latest photos from the jon_og_marteinn_ Instagram account. The widget settings are available at lightwidget.com when signed in as jon_og_marteinn_ from Instagram.

This works when the site has been built but not in development mode. I think it has to do with the <script> tag. I briefly looked up if this widget was available for React but couldn't find anything. Hopefully this can be improved later.