Added a first draft of the simplified logo that fits in a square
I am working on creating a new navigation bar with a logo on the left side of "Jón & Marteinn" in the top right corner. The existing logo is too big to fit. I also prefer to have the text "Jón & Marteinn" in a HTML heading. In the original logo it was part of the image.

This v1 is an initial draft. I may then make another version which would be more exact.