Added the font typeface-lato v0.0.75 as a dependency, etc
Lato is the main font of the Bootstrap theme that is being used. Before this change we loaded it via a <link> element in the <head>. According to the GatsbyJS documentation:

> Hosting your own Google Fonts locally within a project means they won’t have to be fetched over the network when your site loads, increasing your site’s speed index by up to ~300 milliseconds on desktop and 1+ seconds on 3G. It’s also recommended to limit custom font usage to only the essential for performance.

https://www.gatsbyjs.org/docs/recipes/styling-css/#using-google-fonts

I followed the installation instructions in this documentation. I didn't look too deep but it seems that Gatsby or Webpack is able to figure out exactly what font variants of Lato are required. For example at the time of writing the <h1>, <h2> tags use a font weight of 700. Then only that font variant is requested.