Removed FontAwesome icons as a dependency, etc
We were only using one icon from FontAwesome (the Instagram icon). That's too much. I recall that originally I had plans to use more icons but for now I have removed the dependency and replaced the icon with the text "Við á Instagram".

I may add FontAwesome again in the future but hopefully the proper way. I just learned that there is a package called react-fontawesome which utilises SVG under the hood. In conjunction with the default Webpack module bundling configuration that GatsbyJS provides I think that only the icons that are used actually get bundled up.

https://www.gatsbyjs.org/docs/recipes/styling-css/#using-font-awesome

https://github.com/FortAwesome/react-fontawesome