Refactored the CircleCI script to use workflows, etc
Use CircleCI workflows instead of if/else sentences in bash. This simplifies the configuration and removes duplicate steps in deployment. This is going to be important for upcoming changes when we will start to apply Cache-Control HTTP headers to assets.

Moved the AWS CloudFront cache invalidation to a separate step and added a bit of documentation why we are using CloudFront in the first place.