Apply better HTTP cache control headers, use aws-cli v1, v2 is bugged
Apply HTTP cache control headers to certain files to improve performance of the site.

Before this change we were using aws-cli v2 but it seems that when the --metadata-directive REPLACE command is used it applies the wrong MIME type to files. This broke the website in a major way. This is not an issue with aws-cli v1 so I reverted back to it.

In order to be install aws-cli v1 we have to change the CircleCI image. Can't be installed on cimg/node as it's a very minimal. But didn't look too deep into the matter to be honest.

See:
https://www.gatsbyjs.org/docs/caching/