Removed Lato as the primary font
I think Helvetica Neue is good enough. It's available on Windows and macOS. Makes the site faster and eliminates the annoying font-flash on load. I know there are solutions to that but a lower hanging fruit is to simply stop using the font. I may add it or some other custom font again later.