Reorganised the header to be cantered and include subheading
Before this change the header looked like this:

// When wide enough
<logo> J&M           X, Y, Z

// On smaller screens
<logo> J&M
X, Y, Z

After the change it looks like this:

   <logo>
     J&M
   alhliða
    X, Y, Z

What motivated this change was that I wanted to get the subheading "alhliða málningarþjónusta" closer to the main heading "Jón & Marteinn". Doing this paves the way to simplify the site and get rid of the logo that was at the top. It both speeds up the website, looks cleaner and better.