Replaced top of front page with new header navigation elements, etc
Removed the original logo at the top. It has now been replaced with a new header with navigation links. 

Moved the 3x service icons to be below the standing picture of Jón & Marteinn.

Before this change the <h2> heading had a background of a blue/gray painted wall. This has now been removed in favour of just a white background.