Refactored Header and Layout
Wrapped the Header navigation in a Container just like the Footer was. It was looking crammed on mobile.

Split the Layout component into two. Layout wraps all content in a Container and LayoutNoContainer doesn't. Most pages will use Layout but if we need full control LayoutNoContainer is more minimal. At the time of writing it's only used by the front page because it requires a full with. I tried before to add the `fluid` property to the Container but it wasn't enough, it required manually setting 0 padding as well on the sides. Instead I just opted for splitting the Layout component in two.