Feature icons are now black and don't grow bigger on hover
At the time of writing the feature icons can't be clicked. I prefer that nothing has the red colour unless it can be clicked so it is more clear to the user what can be interacted with. Before this change the icons would also grow in size when hovered over. I removed that again because they are just decoration with no interactivity. I opted for overwriting the theme CSS class instead of changing the theme source directly. I want to keep it untouched as much as possible.