Added favicons
Added favicons generated by realfavicongenerator.net. I used the file "simplified logo v1.svg" as a source for the generated assets.

Created the /static folder and added the favicons to it. The file `manifest.webmanifest` is used by Chrome for progressive web apps and when the website is added to the home screen on Android. The file `browserconfig.xml` is used by Windows.

I decided not include `safari-pinned-tab.svg`. I think it's too rare to include.

The file `manifest.webmanifest` was called `site.manifest` in the export. According to the Google docs it can be named anything but by usually it's named `manifest.webmanifest` so I renamed it. I also changed the contents of it a little bit. Added `"start_url": "/"` as a property (it's required according to Google) and changed the theme colour from white to our red theme colour.

Tool used:
https://realfavicongenerator.net/

Original installation instructions:

# Your Favicon Package

This package was generated with [RealFaviconGenerator](https://realfavicongenerator.net/) [v0.16](https://realfavicongenerator.net/change_log#v0.16)

## Install instructions

To install this package:

Extract this package in the root of your web site. If your site is <code>http://www.example.com</code>, you should be able to access a file named <code>http://www.example.com/favicon.ico</code>.

Insert the following code in the `head` section of your pages:

    <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png">
    <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png">
    <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png">
    <link rel="manifest" href="/site.webmanifest">
    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#cf1717">
    <meta name="msapplication-TileColor" content="#2d89ef">
    <meta name="theme-color" content="#ffffff">

*Optional* - Check your favicon with the [favicon checker](https://realfavicongenerator.net/favicon_checker)