Unified navigation links in header and footer into a shared component
Before this change the footer would be left aligned on large screens and in the center on small screens. With this change the navigation links are now in a new component that both the Header and the Footer use. Later down the line I may want to opt to split it into separate components so the footer navigation links can cover a lot more links (something that might not be appropriate for the header).

Deleted the file Header.module.css as the styles were no longer needed. The .header and .header:hover selectors were made for when the logo and title could all be clicked. In an earlier change I made it so that only the image can be clicked so there was no further need for those styles.