Changed /fyrirtækið route to /málarameistari
Renamed "Um Fyrirtækið" and the route /fyrirtækið to "Málarameistari" and /málarameistari. Originally I planned to have those as two separate pages but then I realised the content is virtually the same.

"Málarameistari" is one of they SEO keywords so I hope that it will do better and boost the page a little. That is, I think the route jonogmarteinn.is/málarameistari will do better than jonogmarteinn.is/fyrirtækið.