Headings of all projects are now dynamic
The headings of projects are now dynamic which means that when that project is being viewed as a page it will display <h1>. When the project is being included in some other page it can be specified with a prop what level the heading should be.

Now the project index page has the following headings:

h3: Jón & Marteinn alhliða máningarþjónusta
h1: Verkefni
h2: Vefarastræti
h2: Fosshótel á Hnappavöllum
...

This structure is semantically more correct.