Refactored Málarameistari text, automatic years of experience
Before this change the text description for Málarameistari (on the frontpage and /málarameistari route) were duplicated. With this change the text is now defined in single place so if the text is updated it will stay in sync.

In addition the years of experience is now automatically calculated so we don't have to update it every year. It's a crude calculation that gets incremented on a new year but it will suffice for now.