Fixed <head> tags, now correctly appear in <head> instead of <body>
While reviewing a SEO warning in Lighthouse I noticed that the <head> tags did not appear in the <head> but instead at the top in <body>. After some looking around it seems that specifying `<html lang="is">` in next/head <Head> is not supported. It was supported in react-helmet which was used previously. A mistake I was not careful enough to spot while migrating from react-helmet to next/head.

This also fixes a next-head-count warning on the console. After some further digging it seems this is a known issue[0] with a proposed fix[1].

Since the website will only ever be in Icelandic it is enough to specify a i18n locale setting. That will then set the correct <html> lang attribute.

See:
[0]: https://github.com/vercel/next.js/issues/20924
[1]: https://github.com/vercel/next.js/pull/21080