Generate a robots.txt and sitemap.xml with nextjs-sitemap
A config for next-sitemap to generate a robots.txt and sitemap.xml files for improved SEO. This config is used by the "postbuild" step in package.json. Essentially this config adds a robots.txt file that permits crawling in production but disallows it for any other deployment. All other deployments are dev or staging builds and should not be crawled by search engines to prevent the risk of a duplicate content penalty. A sitemap.xml file is created for production that helps search engines know which routes this site offers. For all non-production deployments no sitemap file is created.