Fixed incorrect sitemap URL
Turns out that the VERCEL_URL is not jonogmarteinn.is when deploying to production as I thought. With these changes the site now uses https://jonogmarteinn.is as a site URL in production.

In addition I added a missing "https://" to other sitemap.xml files. Although it doesn't really matter because they are not generated anyway for non-production deployments.