Added SendGrid mail lib as dependency, removed theme as a dependency
Removed startbootstrap-landing-page as a dependency. It technically wasn't a dependency. I just added it for the record into package.json at the time. At the time I copied the theme file over and have modified it since. The readme still specifies and gives credit to the theme. I removed the dependency to remove a build warning of unmet peer dependencies.

Added SendGrid mail library as a dependency. It's a library that will programmatically send emails via Twilio SendGrid when prospects submit the project request form.