Added initial version of project request form, sends email via SendGrid
Added the first version of the project request form. In it a prospect can fill out a name, phone number, email and a text project description. Once the send button is pressed an API endpoint is invoked (src/pages/api/projectrequest.js) which will send a confirmation email to the prospect and to Jón & Marteinn. Twilio SendGrid is used to deliver the emails.

The projectrequest.js file depends on the environment variable SENDGRID_API_KEY. It is a secret value defined in Vercel.

What remains to be done:
+ Form validation. Right now the forms can be submitted with no values.

+ Place the form properly on the front page. Right now it is located at the top of the page for development convenience.

+ Add a loading spinner while we wait for the API response.

+ Handle error states. For example no internet connectivity and when API response is not a success.

+ Maybe: add Rollbar or Sentry error logging.

+ Maybe: add guards for abuse, for example throttling. Alternatively have a replacement component ready if form needs to be unplugged on a moments notice.