Added loading spinner, etc
Added a square loading spinner. When the submit button is pressed on the project request form it gets replaced with the spinner until the API request has resolved.

Refined the project request confirmation dialog (SubmissionSuccess). Previously it was in a green <Alert> box. Now it's in a <Card> instead. Basically just a box with a white background. It had a built in style for a title and stuff. Looks a little bit better.

Made the props explicitly declared in the SubmissionSuccess component. Hopefully that will make things easier to work with if the props change on the parent for some reason.

After the submit button has been pressed the inputs get disabled while we wait for the API response.

Added a `scrollIntoView()`. After the project request form button has been pressed and the API request has resolved we now scroll to the top of the section. More specifically to the <h1> or <h2> elements with id="verkbeidni". This is a fairly hacky solution but good enough for now I think. I guess the proper way would be to use the `useRef()` hook but I can't be bothered for such a simple case right now.