Navigation links refactor, stop using spacer as a separate <li>
While browsing the site in the text-only browser Lynx I noticed that the navigation looked like this:

* Forsíða
* ⋅
* Verkefni
* ⋅
* Málarameistari
* ⋅
* Verkbeiðni

Refactored it so that the stylistic separator is not its own <li> element but instead comes after each navigation link. Now it looks like this:

* Forsíða⋅
* Verkefni⋅
* Málarameistari⋅
* Verkbeiðni

I have a hunch that this may aid bots that crawl our site.