Minor refactoring, improved copy, etc
Temporarily disabled email sending in the serverless function while it is still in development to prevent exceeding the quota.

The introduction text on the specific Verkbeiðni page now disappears after the form has been submitted. It contained much of the same information so no need to display it twice. Looks better.